+++ CHANGELOG.md
@@ -20,6 +20,34 @@ and this project follows [Semantic Versioning](https://semver.org/).
 - _Nothing yet_
 
 
+## [0.1.11] - 2026-03-09
+
+### Changed
+
+- Add MenuBarStatusPresentation model for prioritized display
+
+## [0.1.10] - 2026-03-09
+
+### Changed
+
+- Add location services support for macOS
+
+## [0.1.9] - 2026-03-09
+
+### Changed
+
+- Add support for local release signing overrides
+
+## [0.1.8] - 2026-03-09
+
+### Added
+
+- Enable WeatherKit integration
+
+### Changed
+
+- Merge branch 'staging'
+
 ## [0.1.7] - 2026-03-09
 
 ### Changed